franky adds compiler flags and gets compiling in pedantic

+++ lib/renderer/unittest/GLRendererTest.h
@@ -80,12 +80,12 @@ void Test()
 	{
 		std::stringstream s;
 		s<<renderer.GetViewPort().GetHeight()<<"=="<<Config::Instance().GetInt("DEFAULT_VIEWPORT_HEIGHT");
-		AssertMsg_Test(renderer.GetViewPort().GetHeight()==Config::Instance().GetInt("DEFAULT_VIEWPORT_HEIGHT"), s.str().c_str());
+		AssertMsg_Test(renderer.GetViewPort().GetHeight()==(size_t)Config::Instance().GetInt("DEFAULT_VIEWPORT_HEIGHT"), s.str().c_str());
 	}
 	{
 		std::stringstream s;
 		s<<renderer.GetViewPort().GetWidth()<<"=="<<Config::Instance().GetInt("DEFAULT_VIEWPORT_WIDTH");
-		AssertMsg_Test(renderer.GetViewPort().GetWidth()==Config::Instance().GetInt("DEFAULT_VIEWPORT_WIDTH"), s.str().c_str());
+		AssertMsg_Test(renderer.GetViewPort().GetWidth()==(size_t)Config::Instance().GetInt("DEFAULT_VIEWPORT_WIDTH"), s.str().c_str());
 	}
 
 	Matrix44 projection = GLRenderer::GetGLProjection();
